feat: add code runner and console formatting utilities
