+++ src/components/CodeBlock.tsx
@@ -1,5 +1,11 @@
 import type { ReactNode } from 'react'
 import { useEffect, useRef, useState } from 'react'
+import {
+  buildTableOutput,
+  formatOutputArgs,
+  formatTraceStack,
+} from '../utils/consoleFormat'
+import { createRunnerWorker, getNodeOnlyReason } from '../utils/codeRunner'
 
 type CodeBlockProps = {
   code: string
@@ -28,285 +34,6 @@ type OutputEntry = {
 }
 
 const javascriptLanguages = new Set(['javascript', 'js'])
-
-const workerSource = `
-const timers = new Map();
-const counts = new Map();
-
-const send = (type, payload) => {
-  self.postMessage({ type, payload });
-};
-
-const safeString = (value) => {
-  if (typeof value === 'string') {
-    return value;
-  }
-  if (value instanceof Error) {
-    return value.stack || value.message || String(value);
-  }
-  try {
-    return JSON.stringify(value, null, 2);
-  } catch (error) {
-    return String(value);
-  }
-};
-
-const console = {
-  log: (...args) => send('log', args),
-  info: (...args) => send('info', args),
-  warn: (...args) => send('warn', args),
-  error: (...args) => send('error', args),
-  table: (...args) => send('table', args),
-  group: (label = 'Group') => send('group', [label]),
-  groupEnd: () => send('groupEnd', []),
-  count: (label = 'default') => {
-    const current = counts.get(label) || 0;
-    const next = current + 1;
-    counts.set(label, next);
-    send('count', [label, next]);
-  },
-  time: (label = 'default') => {
-    timers.set(label, Date.now());
-  },
-  timeEnd: (label = 'default') => {
-    const start = timers.get(label);
-    if (start == null) {
-      send('warn', [
-        'No such label: ' + label,
-      ]);
-      return;
-    }
-    const duration = Date.now() - start;
-    timers.delete(label);
-    send('time', [label, duration]);
-  },
-  trace: (...args) => {
-    const stack = new Error().stack || '';
-    send('trace', [args, stack]);
-  },
-  assert: (condition, ...args) => {
-    if (!condition) {
-      send('assert', args.length ? args : ['Assertion failed']);
-    }
-  },
-  clear: () => send('clear', []),
-};
-
-self.onmessage = (event) => {
-  const { code } = event.data || {};
-  try {
-    const fn = new Function('console', '"use strict";\\n' + code);
-    fn(console);
-    send('done', []);
-  } catch (error) {
-    send('error', [safeString(error)]);
-    send('done', []);
-  }
-};
-`
-
-function createRunnerWorker() {
-  const blob = new Blob([workerSource], { type: 'text/javascript' })
-  const url = URL.createObjectURL(blob)
-  const worker = new Worker(url)
-  URL.revokeObjectURL(url)
-  return worker
-}
-
-function getNodeOnlyReason(code: string) {
-  const matches = [
-    { regex: /\brequire\s*\(/, reason: 'Uses require()' },
-    {
-      regex:
-        /\bimport\s+[^\n]+\s+from\s+['"](fs|path|os|crypto|http|https|net|tls|child_process|zlib|stream|worker_threads|url)['"]/i,
-      reason: 'Uses Node.js modules',
-    },
-    { regex: /\bprocess\./, reason: 'Uses Node.js process' },
-    { regex: /\b__dirname\b|\b__filename\b/, reason: 'Uses Node.js globals' },
-  ]
-
-  for (const match of matches) {
-    if (match.regex.test(code)) {
-      return match.reason
-    }
-  }
-
-  return null
-}
-
-function formatOutputValue(value: unknown) {
-  if (typeof value === 'string') {
-    return value
-  }
-
-  if (value instanceof Error) {
-    return value.stack || value.message || String(value)
-  }
-
-  try {
-    return JSON.stringify(value, null, 2)
-  } catch {
-    return String(value)
-  }
-}
-
-function formatWithPlaceholders(format: string, args: unknown[]) {
-  let result = ''
-  let lastIndex = 0
-  let usedArgs = 0
-  const pattern = /%[sdifoOc%]/g
-
-  format.replace(pattern, (match, offset) => {
-    result += format.slice(lastIndex, offset)
-    lastIndex = offset + match.length
-
-    if (match === '%%') {
-      result += '%'
-      return ''
-    }
-
-    const value = args[usedArgs]
-    usedArgs += 1
-
-    switch (match) {
-      case '%s':
-        result += String(value)
-        break
-      case '%d':
-      case '%i':
-        result += String(Number(value))
-        break
-      case '%f':
-        result += String(Number(value))
-        break
-      case '%o':
-      case '%O':
-        result += formatOutputValue(value)
-        break
-      case '%c':
-        result += ''
-        break
-      default:
-        result += match
-        break
-    }
-
-    return ''
-  })
-
-  result += format.slice(lastIndex)
-  return { text: result, usedArgs }
-}
-
-function formatOutputArgs(args: unknown[]) {
-  if (!args.length) {
-    return ''
-  }
-
-  if (typeof args[0] === 'string') {
-    const { text, usedArgs } = formatWithPlaceholders(args[0], args.slice(1))
-    const remaining = args.slice(1 + usedArgs)
-    const trailing = remaining
-      .map((value) => formatOutputValue(value))
-      .join(' ')
-    return trailing ? `${text} ${trailing}`.trim() : text
-  }
-
-  return args.map((value) => formatOutputValue(value)).join(' ')
-}
-
-function formatTraceStack(rawStack: string) {
-  const lines = rawStack
-    .split('\n')
-    .map((line) => line.trim())
-    .filter(Boolean)
-    .filter(
-      (line) =>
-        !line.startsWith('Error') &&
-        !line.includes('Object.trace') &&
-        !line.includes('self.onmessage'),
-    )
-    .map((line) =>
-      line
-        .replace(/\s*\(eval at self\.onmessage[^)]*\)/g, '')
-        .replace(/\s*\(blob:[^)]*\)/g, ''),
-    )
-
-  return lines.join('\n')
-}
-
-function normalizeTableCell(value: unknown) {
-  const text = formatOutputValue(value)
-  return text.replace(/\s+/g, ' ').trim()
-}
-
-function buildTableOutput(data: unknown) {
-  if (Array.isArray(data)) {
-    if (!data.length) {
-      return '[]'
-    }
-
-    const allKeys = new Set<string>()
-    const isObjectRows = data.every(
-      (row) => row && typeof row === 'object' && !Array.isArray(row),
-    )
-
-    if (isObjectRows) {
-      data.forEach((row) => {
-        Object.keys(row as Record<string, unknown>).forEach((key) => {
-          allKeys.add(key)
-        })
-      })
-
-      const columns = ['(index)', ...Array.from(allKeys)]
-      const rows = data.map((row, index) => {
-        const record = row as Record<string, unknown>
-        return [
-          String(index),
-          ...columns.slice(1).map((key) => normalizeTableCell(record[key])),
-        ]
-      })
-
-      return formatAsciiTable(columns, rows)
-    }
-
-    const columns = ['(index)', 'value']
-    const rows = data.map((row, index) => [
-      String(index),
-      normalizeTableCell(row),
-    ])
-    return formatAsciiTable(columns, rows)
-  }
-
-  if (data && typeof data === 'object') {
-    const entries = Object.entries(data as Record<string, unknown>)
-    if (!entries.length) {
-      return '{}'
-    }
-
-    const columns = ['(index)', 'value']
-    const rows = entries.map(([key, value]) => [key, normalizeTableCell(value)])
-    return formatAsciiTable(columns, rows)
-  }
-
-  return normalizeTableCell(data)
-}
-
-function formatAsciiTable(columns: string[], rows: string[][]) {
-  const widths = columns.map((column, index) => {
-    const rowWidths = rows.map((row) => (row[index] || '').length)
-    return Math.max(column.length, ...rowWidths)
-  })
-
-  const formatRow = (cells: string[]) =>
-    cells.map((cell, index) => cell.padEnd(widths[index])).join(' | ')
-
-  const header = formatRow(columns)
-  const separator = widths.map((width) => '-'.repeat(width)).join('-|-')
-  const body = rows.map((row) => formatRow(row)).join('\n')
-
-  return [header, separator, body].filter(Boolean).join('\n')
-}
 
 async function writeToClipboard(text: string) {
   if (!text) {
